+++ packages/react/helpers/require.ts
@@ -1,4 +1,4 @@
-import React from "react";
+import React from "react"
 
 export const isRequiredChild = (children: React.ReactNode) => React.Children.toArray(children).some(
   (child) => React.isValidElement(child) && child.props.required